refactor: rename worflows to my workflows
