+++ frontend/src/features/workspaces/components/workspaceSettings/SecretsCard.tsx
@@ -17,7 +17,7 @@ import {
 import {
   useAuthenticatedGetRepositorySecrets,
   useAuthenticatedPatchRepositorySecret,
-} from "features/workflows/api";
+} from "features/myWorkflows/api";
 import {
   useState,
   useImperativeHandle,
